requirements.txt
aiogram==3.7.0
SQLAlchemy==2.0.32
asyncpg==0.29.0
python-dotenv==1.0.1
